Alterado arquivo SQL

+++ doc/create_Table.sql
@@ -1,5 +1,5 @@
 create table TS_T_USUARIO(
-ID_USUARIO number(38) primary key,
+ID_USUARIO INTEGER primary key,
 EMAIL VARCHAR2(80),
 SENHA VARCHAR2(150)
 );
@@ -9,5 +9,5 @@ ID_PESSOA number(38) primary key,
 NOME VARCHAR2(20),
 SOBRENOME VARCHAR2(30),
 IDADE NUMBER(2),
-ID_USUARIO NUMBER(38)
+ID_USUARIO integer
 );
